+++ examples/spot/Websocket/SubscribeSpotPrivate.php
@@ -1,27 +1,22 @@
 <?php
 
 use BitMart\Websocket\Spot\WsSpotPrv;
-
 require_once __DIR__ . '/../../../vendor/autoload.php';
-
+// Create Spot Websocket Client
 $ws = new WsSpotPrv([
     'accessKey' => "<your_api_key>",
     'secretKey' => "<your_secret_key>",
     'memo' => "<your_memo>",
-    'xdebug' => false
-]);
-
-// Subscribe Private Channels
-$ws->subscribe(
-    [
-        'op' => "subscribe",
-        'args' => [
-            // Only Support Private Channel
-            "spot/user/order:BTC_USDT",
-        ]
-    ]       ,
-    function ($data) {
+    'callback' => function ($data) {
         echo "-------------------------" . PHP_EOL;
         print_r($data);
-    }
-);
+    },
+]);
+// Login
+$ws->login();
+
+// Subscribe Private Channels
+$ws->send('{"op": "subscribe", "args": ["spot/user/balance:BALANCE_UPDATE"]}');
+
+$ws->send('{"op": "subscribe", "args": ["spot/user/order:BTC_USDT"]}');
+