+++ src/BitMart/Lib/CloudWebsocket.php
@@ -7,7 +7,6 @@ namespace BitMart\Lib;
 use Exception;
 use Ratchet\Client\Connector;
 use Ratchet\Client\WebSocket;
-use Ratchet\RFC6455\Messaging\Frame;
 use Ratchet\RFC6455\Messaging\MessageInterface;
 use React\EventLoop\Loop;
 
@@ -16,40 +15,33 @@ abstract class CloudWebsocket
     protected $url = "";
     protected $xdebug = false;
     protected $useLogin = false;
+    protected $isSpot = false;
+    protected $isClose = false;
     protected $loginParam = false;
+    protected $ready = false;
     protected $reconnectionEnabled = true;
-    protected $reconnectionSubscribeParam = [];
+    protected $reconnectionParam = [];
 
     protected $connection = null;
+    protected $loop = null;
 
-    public function __construct($url, $xdebug, $useLogin)
+    public function __construct($url, $xdebug, $useLogin, $isSpot, $callback, $pong)
     {
         $this->url = $url;
         $this->xdebug = $xdebug;
         $this->useLogin = $useLogin;
+        $this->isSpot = $isSpot;
         $this->reconnectionEnabled = true;
+
+        $this->connection($callback, $pong);
     }
 
-    public function ping()
-    {
-        if ($this->connection) {
-            $this->connection->send(new Frame('', true, Frame::OP_PING));
-        } else {
-            echo("[websockets] ping error");
-        }
-    }
-
-    public function addParam($subscribeParam) {
-        $subscribeParam = json_encode($subscribeParam);
-
-        array_push($this->reconnectionSubscribeParam, $subscribeParam);
-    }
-
-    public function connection($callback): void
+    public function connection($callback, $pong): void
     {
         $reactConnector = new \React\Socket\Connector([
             'timeout' => 10
         ]);
+        $this->ready = false;
         $loop = Loop::get();
         $connector = new Connector($loop, $reactConnector);
 
@@ -58,23 +50,15 @@ abstract class CloudWebsocket
         }
 
         $connector($this->url)
-            ->then(function(WebSocket $conn) use ($loop, $callback) {
+            ->then(function(WebSocket $conn) use ($loop, $callback, $pong) {
                 $this->connection = $conn;
+
 
                 if ($this->xdebug) {
                     echo("[websockets] Connected " . $this->url . PHP_EOL);
                 }
 
-                if ($this->useLogin) {
-                    if ($this->xdebug) {
-                        echo("[websockets] Send:" . $this->loginParam . PHP_EOL);
-                    }
-
-                    $conn->send($this->loginParam);
-
-                }
-
-                $conn->on('message', function(MessageInterface $msg) use ($loop, $callback, $conn) {
+                $conn->on('message', function(MessageInterface $msg) use ($loop, $callback, $pong, $conn) {
 
                     if($msg->isBinary()) {
                         $data = gzinflate($msg->getPayload());
@@ -82,51 +66,76 @@ abstract class CloudWebsocket
                         $data = $msg->getPayload();
                     }
 
-                    if ($data === 'pong') {
-                        return;
-                    }
 
                     if ($this->xdebug) {
                         echo("[websockets] Recv:" . $data . PHP_EOL);
                     }
 
-                    $data = json_decode($data, true);
-                    call_user_func_array($callback, [$data]);
+                    if ($data === 'pong') {
+                        if($pong) {
+                            call_user_func_array($pong, [$data]);
+                        }
+                        return;
+                    }
 
-                    if (isset($data['event']) && $data['event'] === 'login' && isset($data['errorCode'])) {
+                    $data_json = json_decode($data, true);
+                    if($callback) {
+                        call_user_func_array($callback, [$data_json]);
+                    }
+
+                    if (isset($data_json['event']) && $data_json['event'] === 'login' && isset($data_json['errorCode'])) {
                         $this->reconnectionEnabled = false;
+                        $this->isClose = true;
+                        $this->ready = false;
                         $conn->close($reason=$data);
                         $loop->stop();
 
-                    } elseif (isset($data['action']) && $data['action'] === 'access' && isset($data['error'])) {
+                    } elseif (isset($data_json['action']) && $data_json['action'] === 'access' && isset($data_json['error'])) {
                         $this->reconnectionEnabled = false;
+                        $this->isClose = true;
+                        $this->ready = false;
                         $conn->close($reason=$data);
                         $loop->stop();
                     }
                 });
 
-                $conn->on('close', function($reason) use ($callback, $loop) {
+                $conn->on('close', function($reason) use ($callback, $pong, $loop) {
                     if ($this->xdebug) {
-                        echo "[websockets] Connection closed." . print_r($reason) . PHP_EOL;
+                        echo "[websockets] Connection closed." . print_r($reason, true) . PHP_EOL;
+                    }
+
+                    if ($this->isClose) {
+                        echo "[websockets] Client closed." . PHP_EOL;
+                        return ;
                     }
 
                     if ($this->reconnectionEnabled) {
                         echo "[websockets] Reconnecting to " . $this->url . PHP_EOL;
 
-                        $loop->addTimer(2, function () use ($callback) {
-                            $this->connection($callback);
+                        $loop->addTimer(2, function () use ($callback, $pong) {
+                            $this->connection($callback, $pong);
                         });
                     }
                 });
 
-                $loop->addTimer(2, function () use ($conn) {
-                    foreach ($this->reconnectionSubscribeParam as $key => $value) {
-                        if ($this->xdebug) {
-                            echo("[websockets] Send:" . $value . PHP_EOL);
+
+                $loop->addTimer(2, function () use ($loop, $conn) {
+
+                    if ($this->useLogin) {
+                        $conn->send($this->loginParam);
+                    }
+
+                    $loop->addTimer(2, function () use ($conn) {
+                        foreach ($this->reconnectionParam as $key => $value) {
+                            if ($this->xdebug) {
+                                echo("[websockets] Send:" . $value . PHP_EOL);
+                            }
+
+                            $conn->send($value);
                         }
 
-                        $conn->send($value);
-                    }
+                        $this->ready = true;
+                    });
                 });
 
 
@@ -136,10 +145,71 @@ abstract class CloudWebsocket
             });
 
         $loop->addPeriodicTimer(10, function () {
-            $this->ping();
+            $this->keepalive();
         });
     }
 
+    public function send(string $message)
+    {
+        if (in_array($message, $this->reconnectionParam, true)) {
+            if ($this->xdebug) {
+                echo "[websockets] Message already sent: " . $message . PHP_EOL;
+            }
+            return;
+        }
+
+        $this->reconnectionParam[] = $message;
+
+        if ($this->ready) {
+            if ($this->xdebug) {
+                echo "[websockets] Send." . $message . PHP_EOL;
+            }
+            $this->connection->send($message);
+        }
+    }
+
+    public function login()
+    {
+        $this->useLogin = true;
+        $timestamp = round(microtime(true) * 1000);
+        $sign = CloudUtil::signature($timestamp, "bitmart.WebSocket", $this->cloudConfig);
+
+        if ($this->isSpot) {
+            $this->loginParam = json_encode([
+                'op' => "login",
+                'args' => [$this->cloudConfig->accessKey, $timestamp, $sign]
+            ]);
+        } else {
+            $this->loginParam = json_encode([
+                'action' => "access",
+                'args' => [$this->cloudConfig->accessKey, $timestamp.'', $sign, 'web']
+            ]);
+        }
+    }
+
+    public function keepalive()
+    {
+        if ($this->isClose) {
+            echo "[websockets] Client closed, stop send ping." . PHP_EOL;
+            return ;
+        }
+
+        if ($this->connection) {
+            if ($this->isSpot) {
+                $this->connection->send('ping');
+
+            } else {
+                $this->connection->send('{"action":"ping"}');
+            }
+        } else {
+            echo("[websockets] ping error");
+        }
+    }
+
+    public function stop()
+    {
+        $this->isClose = true;
+    }
 
 
 }